fix: fetch data from inside sidebar component
